requirements.txt
invoke==2.2.0
loguru==0.7.3
Markdown==3.7
matplotlib==3.10.0
Pillow==11.1.0
pydantic==2.10.5
pytorch_lightning==2.5.0.post0
torch==2.5.1
torchmetrics==1.6.1
torchvision==0.20.1
typer==0.15.1
fastapi==0.115.6
uvicorn==0.34.0
ruff==0.8.6
hydra-core==1.3.2
wandb==0.19.4
pytest==8.3.4
coverage==7.6.10
pre-commit==4.1.0
onnxruntime==1.20.1
onnx==1.17.0
streamlit==1.41.1
google-cloud-run==0.10.14
google-cloud-storage==2.19.0
python-multipart==0.0.20
prometheus_client==0.21.1
